Test: 1. delete the task if it exists 2. do not invoke delete if the task not found 3. save the correct task when it is added to tasks 4. update the correct task data when updateTask() method is invoked

+++ src/main/java/io/github/sfakaly/service/TaskService.java
@@ -1,11 +1,11 @@
 package io.github.sfakaly.service;
 
 import io.github.sfakaly.exceptions.EmptyListException;
+import io.github.sfakaly.exceptions.TaskNotFoundException;
 import io.github.sfakaly.model.Task;
 import io.github.sfakaly.repository.JsonTaskRepository;
 import lombok.RequiredArgsConstructor;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,8 +13,8 @@ import java.util.List;
 public class TaskService {
     private final JsonTaskRepository repository;
 
-    public void addTask(String title, boolean isDone, LocalDateTime createdAt) {
-        Task task = new Task(0, title, createdAt, isDone);
+    public void addTask(String title) {
+        Task task = new Task(title);
         repository.saveTask(task);
     }
 
@@ -23,7 +23,8 @@ public class TaskService {
     }
 
     public void deleteTask(int id) {
-        repository.findById(id);
+        Task task = repository.findById(id);
+        if (task == null) throw new TaskNotFoundException("Задачи с ID " + id + " не существует.");
         repository.deleteTaskById(id);
     }
 
@@ -33,7 +34,7 @@ public class TaskService {
 
     public void updateTask(Task task, String newTitle, boolean newIsDone) {
         if (!newTitle.isBlank()) task.setTitle(newTitle);
-        task.setIsDone(newIsDone);
+        task.setCompleted(newIsDone);
         repository.updateStorage();
     }
